Removal of old Api::Json::MapsController
