+++ Makefile
@@ -254,7 +254,6 @@ SPEC_HELPER_MIN_SPECS = \
 	spec/models/carto/user_migration_spec.rb \
 	spec/models/table_registrar_spec.rb \
 	spec/requests/admin/organization_users_controller_spec.rb \
-	spec/requests/api/json/maps_controller_spec.rb \
 	spec/requests/carto/admin/mobile_apps_controller_spec.rb \
 	spec/requests/carto/builder/datasets_controller_spec.rb \
 	spec/requests/carto/api/analyses_controller_spec.rb \
