--- spec/requests/api/json/maps_controller_shared_examples.rb
@@ -1,5 +1,0 @@
-# encoding: utf-8
-
-shared_examples_for "maps controllers" do
-
-end
